backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cloudinary==1.26.0
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.2
oauthlib==3.1.1
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
